增加对值类型和JToken类型的列表读取

+++ MyStaging/Helpers/UpdateBuilder.cs
@@ -21,7 +21,7 @@ namespace MyStaging.Helpers
         /// <param name="field"></param>
         /// <param name="value"></param>
         /// <returns></returns>
-        public UpdateBuilder<T> SetIncrement(string field, double value)
+        public UpdateBuilder<T> SetIncrement(string field, decimal value)
         {
             setList.Add($"{field}={field} + {value}");
             return this;
